+++ vite.config.ts
@@ -1,7 +1,15 @@
 import { defineConfig } from 'vite';
 import { crx } from '@crxjs/vite-plugin';
 import manifest from './manifest.json';
+import { resolve } from 'path';
 
 export default defineConfig({
   plugins: [crx({ manifest })],
+  build: {
+    rollupOptions: {
+      input: {
+        onboarding: resolve(__dirname, 'src/onboarding/index.html')
+      }
+    }
+  }
 });